// src/model/enum/product-sort-option.enum.ts
export enum ProductSortOption {
  PRICE_LOW_TO_HIGH = 'PRICE_LOW_TO_HIGH',
  PRICE_HIGH_TO_LOW = 'PRICE_HIGH_TO_LOW',
  NAME_A_TO_Z = 'NAME_A_TO_Z',
  NAME_Z_TO_A = 'NAME_Z_TO_A',
  CREATION_DATE_HIGH_TO_LOW = 'CREATION_DATE_HIGH_TO_LOW',
  CREATION_DATE_LOW_TO_HIGH = 'CREATION_DATE_LOW_TO_HIGH',
  // MOST_SOLD,
}